use is sub type
