+++ modules/ismp/pallet/call-decompressor/Cargo.toml
@@ -13,12 +13,13 @@ sp-runtime = { workspace = true }
 sp-core = { workspace = true }
 sp-std = { workspace = true }
 sp-io = { workspace = true }
-pallet-ismp = { path = "../ismp/pallet", default-features = false }
+pallet-ismp = { path = "..", default-features = false }
 pallet-ismp-relayer = {path = "../relayer", default-features = false }
+as-any = "0.3.1"
 
 
 # polytope labs
-ismp = { package = "ismp", path = "../ismp/core", default-features = false }
+ismp = { package = "ismp", path = "../../core", default-features = false }
 
 # crates.io
 codec = { package = "parity-scale-codec", version = "3.1.3", default-features = false }
